1- Rename the module from aiXtend to aixplain 2- Class Metric now contain __repr__ function 3- the function create_benchmark_from_id in the class BenchmarkFactory is renamed to create_asset_from_id 4- paths in the the tests files are modified 5- fixing the model testing function which was using the old link
