+++ README.md
@@ -1,8 +1,8 @@
 <img src="docs/assets/aixplain-brandmark-common.png" alt="aiXplain logo" title="aiXplain" align="right" height="132" width="85"/>
 
-# aiXtend
+# aixplain
 
-aiXtend is a software development kit (SDK) for the [aiXplain](https://aixplain.com/) platform. With aiXtend, developers can quickly and easily:
+aixplain is a software development kit (SDK) for the [aiXplain](https://aixplain.com/) platform. With aiXtend, developers can quickly and easily:
 
 - [Discover](https://aixplain.com/platform/discovery/) aiXplain’s ever-expanding catalog of 35,000+ ready-to-use AI models and utilize them.
 - [Benchmark](https://aixplain.com/platform/benchmark/) AI systems by choosing models, datasets and metrics.
@@ -19,11 +19,11 @@ aiXtend is a software development kit (SDK) for the [aiXplain](https://aixplain.
 ### Installation
 To install simply,
 ```bash
-pip install aixtend
+pip install aixplain
 ```
 
 ###  API Key Setup
-Before you can use the aiXtend SDK, you'll need to obtain an API key from our platform. For details refer this [Team API Key Guide](docs/user/api_setup.md).
+Before you can use the aixplain SDK, you'll need to obtain an API key from our platform. For details refer this [Team API Key Guide](docs/user/api_setup.md).
 
 Once you get the API key, you'll  need to add this API key as an environment variable on your system.
 #### Linux or macOS
@@ -36,10 +36,10 @@ set TEAM_API_KEY=YOUR_API_KEY
 ```
 ## Usage
 
-Let’s see how we can use aiXtend to run a machine translation model.
+Let’s see how we can use aixplain to run a machine translation model.
 
 ```python
-from aixtend.factories.model_factory  import ModelFactory
+from aixplain.factories.model_factory  import ModelFactory
 model = ModelFactory.create_asset_from_id("61b27086c45ecd3c10d0608c") # Got the ID of an MT model from on our platform
 translation = model.run("This is a sample text")
 ```
