+++ aixplain/modules/dataset.py
@@ -16,23 +16,24 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 Author: Duraikrishna Selvaraju, Thiago Castro Ferreira, Shreyas Sharma and Lucas Pavanelli
-Date: October 25th 2022
+Date: October 28th 2022
 Description:
-    Metric Class
+    Datasets Class
 """
-
 from typing import List
-from aixtend.utils.file_utils import _request_with_retry
+from aixplain.utils.file_utils import _request_with_retry
+from aixplain.utils import config
 
-class Metric:
+
+class Dataset:
     def __init__(self, id:str, name:str, description:str, **additional_info) -> None:
-        """Create a Metric with the necessary information
+        """Create a Dataset with the necessary information
 
         Args:
-            id (str): ID of the Metric
-            name (str): Name of the Metric
-            description (str): Description of the Metric
-            **additional_info: Any additional Metric info to be saved
+            id (str): ID of the Dataset
+            name (str): Name of the Dataset
+            description (str): Description of the Dataset
+            **additional_info: Any additional dataset info to be saved
         """
         self.id = id
         self.name = name
@@ -40,14 +41,9 @@ class Metric:
         self.additional_info = additional_info
 
     def get_asset_info(self) -> dict:
-        """Get the Metric info as a Dictionary
+        """Get the dataset info as a Dictionary
 
         Returns:
-            dict: Metric Information
+            dict: Dataset Information
         """
         return self.__dict__
-
-    
-
-    
-
