+++ docs/samples/subtitle_generator/subtitle_generator.py
@@ -35,7 +35,7 @@ import logging
 import requests
 from tqdm import tqdm
 
-from aixtend.factories.pipeline_factory import PipelineFactory
+from aixplain.factories.pipeline_factory import PipelineFactory
 
 
 def generate_srt(segments: list, write_file: str = None):
