+++ setup.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 """
-Copyright 2022 The aiXtend authors
+Copyright 2022 The aixplain authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -54,7 +54,7 @@ test_requirements = [
 about = {}
 here = os.path.abspath(os.path.dirname(__file__))
 
-with open(os.path.join(here, "aixtend", "__version__.py"), "r") as f:
+with open(os.path.join(here, "aixplain", "__version__.py"), "r") as f:
     exec(f.read(), about)
 
 with open("README.md", "r") as f:
@@ -70,7 +70,7 @@ setup(
     author_email=about["__author_email__"],
     url=about["__url__"],
     packages=find_packages(exclude=["test"]),
-    package_dir={"aixtend": "aixtend"},
+    package_dir={"aixplain": "aixplain"},
     package_data={"": ["LICENSE"]},
     include_package_data=True,
     python_requires=">=3.5, <4",
