+++ tests/benchmark_test.py
@@ -5,9 +5,9 @@ import pytest
 import requests_mock
 from pathlib import Path
 import pandas as pd
-from aixtend.utils import config
+from aixplain.utils import config
 
-from aixtend.factories.benchmark_factory import BenchmarkFactory, ModelFactory, DatasetFactory, MetricFactory
+from aixplain.factories.benchmark_factory import BenchmarkFactory, ModelFactory, DatasetFactory, MetricFactory
 
 FIXED_HEADER = {
     'Authorization': f"Token {config.TEAM_API_KEY}",
@@ -32,21 +32,21 @@ def __mock_benchmark_create_dependecies(mock, asset_id):
     }
     for asset_name in asset_list:
         url = url_dict[asset_name]
-        with open(Path("tests\mock_responses\get_asset_info_responses.json")) as f:
+        with open(Path("tests/mock_responses/get_asset_info_responses.json")) as f:
             mock_json = json.load(f)[asset_name]
         mock.get( url, headers=FIXED_HEADER, json=mock_json)
-        with open(Path("tests\mock_responses\get_asset_info_responses.json")) as f:
+        with open(Path("tests/mock_responses/get_asset_info_responses.json")) as f:
             mock_json = json.load(f)['benchmarkJob']
         url_benchmarkJobs = f"{config.BENCHMARKS_BACKEND_URL}/sdk/benchmarks/{asset_id}/jobs"
         mock.get( url_benchmarkJobs, headers=FIXED_HEADER, json=[mock_json])
 
         url_benchmark = f"{config.BENCHMARKS_BACKEND_URL}/sdk/benchmarks/{asset_id}"
-        with open(Path("tests\mock_responses\get_asset_info_responses.json")) as f:
+        with open(Path("tests/mock_responses/get_asset_info_responses.json")) as f:
             mock_json = json.load(f)['benchmark']
         mock.get( url_benchmark, headers=FIXED_HEADER, json=mock_json)
 
         url_benchmarkJob = f"{config.BENCHMARKS_BACKEND_URL}/sdk/benchmarks/jobs/{asset_id}"
-        with open(Path("tests\mock_responses\get_asset_info_responses.json")) as f:
+        with open(Path("tests/mock_responses/get_asset_info_responses.json")) as f:
             mock_json = json.load(f)['benchmarkJob']
         mock.get( url_benchmarkJob, headers=FIXED_HEADER, json=mock_json)
 
@@ -81,7 +81,7 @@ def test_create_benchmark_with_assets():
             benchmark_creation_params.append([asset])
 
         url = f"{config.BENCHMARKS_BACKEND_URL}/sdk/benchmarks"
-        with open(Path("tests\mock_responses\get_asset_info_responses.json")) as f:
+        with open(Path("tests/mock_responses/get_asset_info_responses.json")) as f:
             mock_json = json.load(f)['benchmark']
         mock.post(url, headers=FIXED_HEADER, json=mock_json)
         benchmark = BenchmarkFactory.create_benchmark(*benchmark_creation_params)
@@ -93,7 +93,7 @@ def test_start_benchmark_job():
     with requests_mock.Mocker() as mock:
         __mock_benchmark_create_dependecies(mock, asset_id)
         url = f"{config.BENCHMARKS_BACKEND_URL}/sdk/benchmarks/{asset_id}/start"
-        with open(Path("tests\mock_responses\get_asset_info_responses.json")) as f:
+        with open(Path("tests/mock_responses/get_asset_info_responses.json")) as f:
             mock_json = json.load(f)['benchmarkJob']
         mock.post(url, headers=FIXED_HEADER, json=mock_json)
         benchmark = BenchmarkFactory.create_benchmark_from_id(asset_id)
@@ -107,9 +107,9 @@ def test_download_benchmark_results():
     with requests_mock.Mocker() as mock:
         __mock_benchmark_create_dependecies(mock, asset_id)
         benchmarkJob = BenchmarkFactory.create_benchmark_job_from_id(asset_id)
-        with open(Path("tests\mock_responses\get_asset_info_responses.json")) as f:
+        with open(Path("tests/mock_responses/get_asset_info_responses.json")) as f:
             download_url = json.load(f)['benchmarkJob']['reportUrl']
-        dummy_result_path = Path("tests\mock_responses\dummy_result.csv")
+        dummy_result_path = Path("tests/mock_responses/dummy_result.csv")
         with open(dummy_result_path, "rb") as f:
             dummy_content = f.read()
         mock.get(download_url, headers=FIXED_HEADER, content=dummy_content)
